# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TuTramiteFacil API",
        "description": "API REST para gestión de solicitantes y solicitudes de ayudas sociales (DDD + Clean Architecture).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8010",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/solicitudes": {
            "get": {
                "tags": [
                    "Solicitudes"
                ],
                "summary": "Listar todas las solicitudes",
                "description": "GET /api/solicitudes — público.",
                "operationId": "af2fa1a2817177309b1fff464a0c6b27",
                "responses": {
                    "200": {
                        "description": "Lista de solicitudes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Solicitudes"
                ],
                "summary": "Crear una nueva solicitud",
                "description": "Al crear la solicitud se encola ProcesarSolicitudJob en Horizon para procesamiento asíncrono.",
                "operationId": "f322569936bf9baef1864569414a2395",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "solicitante_id",
                                    "tipo_ayuda",
                                    "fecha_solicitud",
                                    "importe_estimado"
                                ],
                                "properties": {
                                    "solicitante_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "uuid-del-solicitante"
                                    },
                                    "tipo_ayuda": {
                                        "type": "string",
                                        "example": "Alquiler",
                                        "enum": [
                                            "Alquiler",
                                            "ChequeBebe",
                                            "IngresoMinimoVital",
                                            "BonoCulturalJoven"
                                        ]
                                    },
                                    "fecha_solicitud": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-15"
                                    },
                                    "fecha_resolucion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "importe_estimado": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "Pendiente",
                                        "enum": [
                                            "Pendiente",
                                            "EnRevision",
                                            "Concedida",
                                            "Denegada"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Solicitud creada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/solicitudes/{solicitud}": {
            "get": {
                "tags": [
                    "Solicitudes"
                ],
                "summary": "Ver detalle de una solicitud",
                "description": "GET /api/solicitudes/{solicitud} — público.",
                "operationId": "93ac7390bbeb5dcebfac87026a2527c6",
                "parameters": [
                    {
                        "name": "solicitud",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos de la solicitud"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Solicitudes"
                ],
                "summary": "Actualizar una solicitud",
                "description": "Si el campo estado cambia, se encola ProcesarSolicitudJob para notificar el nuevo estado.",
                "operationId": "2546a9f29f26e7e6b22138da3775ffb2",
                "parameters": [
                    {
                        "name": "solicitud",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "solicitante_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "tipo_ayuda": {
                                        "type": "string",
                                        "enum": [
                                            "Alquiler",
                                            "ChequeBebe",
                                            "IngresoMinimoVital",
                                            "BonoCulturalJoven"
                                        ]
                                    },
                                    "fecha_solicitud": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "fecha_resolucion": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "importe_estimado": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "Pendiente",
                                            "EnRevision",
                                            "Concedida",
                                            "Denegada"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Solicitud actualizada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Solicitudes"
                ],
                "summary": "Eliminar una solicitud",
                "description": "DELETE /api/solicitudes/{solicitud} — requiere JWT.",
                "operationId": "428f2a256038f899e6864069bf9f029d",
                "parameters": [
                    {
                        "name": "solicitud",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminada correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}